created seperated component TaskInProgress
